requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
dj-database-url==3.1.0
Django==6.0.1
django-allauth==65.13.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
razorpay==2.0.0
requests==2.32.5
sqlparse==0.5.5
stripe==14.1.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
